fixup: exclude self from assets list
When looking at the assets tied to a DigitalAsset, we get a strange
exception unless we skip the record that corresponds to the DigitalAsset
itself. It's strange that we see a record like this when digital_assets
are supposed to have their own table for storage, but there it is.